+++ js/util.js
@@ -1,0 +1,22 @@
+// Функция для генерации случайного числа
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// Функция для генерации ID
+
+function generateId() {
+  let id = 0;
+  return function () {
+    return id++;
+  };
+}
+
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, generateId, getRandomArrayElement};
